Access

+++ app/Http/Controllers/ProjectController.php
@@ -15,7 +15,10 @@ class ProjectController extends Controller
     public function index()
     {
         try {
+            if(Auth()->user()->hasRoleAdmin())
             return view('project.index', ['projects' => Project::paginate(10)]);
+            return  redirect()->route("task.index")->with('status', '************ YOU DO NOT HAVE ACCESS TO PROJECTS ! *************');
+
         } catch (Exception $ex) {
             Log::critical("index error project".$ex->getMessage());
             abort(500);
@@ -28,7 +31,10 @@ class ProjectController extends Controller
     public function create()
     {
         try {
+            if(Auth()->user()->hasRoleAdmin())
             return view('project.create');
+            return  redirect()->route("task.index")->with('status', '************ YOU DO NOT HAVE ACCESS TO CREATE PROJECT ! *************');
+
         } catch (Exception $ex) {
             Log::critical("create error project".$ex->getMessage());
             abort(500);
@@ -64,7 +70,10 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {   try {
+        if(Auth()->user()->hasRoleAdmin())
             return view('project.update', ['project' =>  $project]);
+            return  redirect()->route("task.index")->with('status', '************ YOU DO NOT HAVE ACCESS TO UPDATE PROJECT ! *************');
+
         } catch (Exception $ex) {
             Log::critical("edit error project".$ex->getMessage());
             abort(500);
@@ -101,8 +110,13 @@ class ProjectController extends Controller
     public function destroy(Project $project)
     {
         try {
+            if (Auth()->user()->hasRoleAdmin())
+            {
             $project->delete();
             return redirect()->route('project.index')->with('status', 'Project has been successfully suppressed.');
+            }
+            return  redirect()->route("task.index")->with('status', '************ YOU DO NOT HAVE ACCESS TO DELETE PROJECT ! *************');
+
         } catch (Exception $ex) {
             Log::critical("destroy error project".$ex->getMessage());
             abort(500);
